+++ test/data/sandbox/browser_test.multiple.js
@@ -1,5 +1,5 @@
 Feature('print browser');
 
-Scenario('print browser info', (I) => {
+Scenario('print browser info', ({ I }) => {
   I.printBrowser();
 });