+++ test/data/sandbox/workers/failing_test.worker.js
@@ -4,7 +4,7 @@ Before(() => {
   throw new Error('worker has failed');
 });
 
-Scenario('should not be executed', (I) => {
+Scenario('should not be executed', ({ I }) => {
   I.say('Hello Workers');
   I.seeThisIsWorker();
 });